Add an additional info field to contact triage and display below form
Fix #1363

To address requirement to have switchboard on the main contact triage, but
below the form.
Adds an extra formatted text field, which is hidden on the node view.
Amend the contact triage block to show below the form.
Add some extra styling to add some margin between the form and the
additional information.